add_file

+++ inc/Search_server.h
@@ -12,7 +12,7 @@ public:
   std::vector<std::vector<RelativeIndex>> search(const std::vector<std::string>& queries_input) const;
 private:
   InvertedIndex& _index;
-  int GetResponsesLimit()const;
+  size_t GetResponsesLimit()const;
   std::map<size_t, size_t> GetABSRelevance(const std::string& request)const;
-  std::vector<RelativeIndex> GetRelRelevance(const std::map<size_t, size_t>& abs_relevance, int responces_limit) const;
+  std::vector<RelativeIndex> GetRelRelevance(const std::map<size_t, size_t>& abs_relevance, size_t responces_limit) const;
 };
